Pull apart scheduled/mod.rs
I pulled this file into separate parts:
1. graph.rs, which contains all the atomic operations we need and internal methods.
2. graph_ext.rs, which contains things that can be built out of those atomic
   operations, like the helper methods for adding operators.
3. tests.rs, which includes the test previously at the bottom of mod.rs.
4. context.rs, which holds the stuff around a `Context`.
5. reactor.rs, which holds stuff around a `Reactor`.
